services/UberPendragon/backend/requirements.txt
APScheduler==3.10.1
cryptography==39.0.2
Flask==2.3.2
Flask-Cors==4.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
pycryptodome==3.16.0
fastecdsa==2.2.3